+++ doc/user/content/release-notes.md
@@ -128,6 +128,8 @@ changes that have not yet been documented.
 
 - Add the [`pg_catalog.pg_views`](/sql/system-catalog#pg-catalog) PostgreSQL compatibility view.
 
+- Add the [`information_schema.columns`](/sql/system-catalog#information-schema) compatibility view.
+
 {{< comment >}}
 Only add new release notes above this line.